+++ test/urlencoded.js
@@ -211,6 +211,114 @@ describe('bodyParser.urlencoded()', function(){
     })
   })
 
+  describe('with parameterLimit option', function () {
+    describe('with extended: false', function () {
+      it('should reject 0', function () {
+        assert.throws(createServer.bind(null, { extended: false, parameterLimit: 0 }), /option parameterLimit/)
+      })
+
+      it('should reject string', function () {
+        assert.throws(createServer.bind(null, { extended: false, parameterLimit: 'beep' }), /option parameterLimit/)
+      })
+
+      it('should 415 if over limit', function (done) {
+        request(createServer({ extended: false, parameterLimit: 10 }))
+        .post('/')
+        .set('Content-Type', 'application/x-www-form-urlencoded')
+        .send(createManyParams(11))
+        .expect(413, /too many parameters/, done)
+      })
+
+      it('should work when at the limit', function (done) {
+        request(createServer({ extended: false, parameterLimit: 10 }))
+        .post('/')
+        .set('Content-Type', 'application/x-www-form-urlencoded')
+        .send(createManyParams(10))
+        .expect(expectKeyCount(10))
+        .expect(200, done)
+      })
+
+      it('should work if number is floating point', function (done) {
+        request(createServer({ extended: false, parameterLimit: 10.1 }))
+        .post('/')
+        .set('Content-Type', 'application/x-www-form-urlencoded')
+        .send(createManyParams(11))
+        .expect(413, /too many parameters/, done)
+      })
+
+      it('should work with large limit', function (done) {
+        request(createServer({ extended: false, parameterLimit: 5000 }))
+        .post('/')
+        .set('Content-Type', 'application/x-www-form-urlencoded')
+        .send(createManyParams(5000))
+        .expect(expectKeyCount(5000))
+        .expect(200, done)
+      })
+
+      it('should work with Infinity limit', function (done) {
+        request(createServer({ extended: false, parameterLimit: Infinity }))
+        .post('/')
+        .set('Content-Type', 'application/x-www-form-urlencoded')
+        .send(createManyParams(10000))
+        .expect(expectKeyCount(10000))
+        .expect(200, done)
+      })
+    })
+
+    describe('with extended: true', function () {
+      it('should reject 0', function () {
+        assert.throws(createServer.bind(null, { extended: true, parameterLimit: 0 }), /option parameterLimit/)
+      })
+
+      it('should reject string', function () {
+        assert.throws(createServer.bind(null, { extended: true, parameterLimit: 'beep' }), /option parameterLimit/)
+      })
+
+      it('should 415 if over limit', function (done) {
+        request(createServer({ extended: true, parameterLimit: 10 }))
+        .post('/')
+        .set('Content-Type', 'application/x-www-form-urlencoded')
+        .send(createManyParams(11))
+        .expect(413, /too many parameters/, done)
+      })
+
+      it('should work when at the limit', function (done) {
+        request(createServer({ extended: true, parameterLimit: 10 }))
+        .post('/')
+        .set('Content-Type', 'application/x-www-form-urlencoded')
+        .send(createManyParams(10))
+        .expect(expectKeyCount(10))
+        .expect(200, done)
+      })
+
+      it('should work if number is floating point', function (done) {
+        request(createServer({ extended: true, parameterLimit: 10.1 }))
+        .post('/')
+        .set('Content-Type', 'application/x-www-form-urlencoded')
+        .send(createManyParams(11))
+        .expect(413, /too many parameters/, done)
+      })
+
+      it('should work with large limit', function (done) {
+        request(createServer({ extended: true, parameterLimit: 5000 }))
+        .post('/')
+        .set('Content-Type', 'application/x-www-form-urlencoded')
+        .send(createManyParams(5000))
+        .expect(expectKeyCount(5000))
+        .expect(200, done)
+      })
+
+      it('should work with Infinity limit', function (done) {
+        request(createServer({ extended: true, parameterLimit: Infinity }))
+        .post('/')
+        .set('Content-Type', 'application/x-www-form-urlencoded')
+        .send(createManyParams(10000))
+        .expect(expectKeyCount(10000))
+        .expect(200, done)
+      })
+    })
+  })
+
   describe('with type option', function(){
     var server;
     before(function(){
@@ -371,6 +479,23 @@ describe('bodyParser.urlencoded()', function(){
   })
 })
 
+function createManyParams(count) {
+  var str = ''
+
+  if (count === 0) {
+    return str
+  }
+
+  str += '0=0'
+
+  for (var i = 1; i < count; i++) {
+    var n = i.toString(36)
+    str += '&' + n + '=' + n
+  }
+
+  return str
+}
+
 function createServer(opts){
   var _bodyParser = bodyParser.urlencoded(opts)
 
@@ -380,4 +505,10 @@ function createServer(opts){
       res.end(err ? err.message : JSON.stringify(req.body));
     })
   })
+}
+
+function expectKeyCount(count) {
+  return function (res) {
+    assert.equal(Object.keys(JSON.parse(res.text)).length, count)
+  }
 }